requirements.txt
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0
pydub>=0.25.1
typer>=0.12.0
rich>=13.7.0
pyyaml>=6.0
tqdm>=4.66.0